# Cache Abbreviation entries after materializing them to speedup MCCAS replay