When looking at a trace of MCCAS replay, we noticed that a large chunk of time is spent in the function `visitDIEAttrs`

The reason for that is because, we materialize abbreviations every time we see a new DIE, which leads to decoding the same ULEB128 values that represent the form and attribute of a DIE for every DIE in the debug info section. We can materialize them once and cache them, which would reduce the time spent and give us some speedup.

We see some speedup as well:

MCCAS With Abbrev Caching
ninja clang  150.46s user 48.35s system 536% cpu 37.071 total

ninja clang  150.43s user 47.33s system 561% cpu 35.240 total

ninja clang  150.78s user 47.23s system 565% cpu 34.987 total

MCCAS Without Abbrev Caching
ninja clang  193.36s user 48.17s system 557% cpu 43.320 total

ninja clang  190.28s user 46.98s system 595% cpu 39.867 total

ninja clang  191.55s user 46.95s system 602% cpu 39.602 total